https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=832265
The switch to using -fstack-protector-strong in stretch has exposed a
stack-smashing bug.

The problem affects 64-bit platforms, as the code is assuming
sizeof(pid_t) == sizeof(long), yet on x86_64:

sizeof(pid_t) == 4
sizeof(long) == 8

The relevant code is in suckutils.c, do_lockfile():

int do_lockfile(PMaster master) {
...
    pid_t pid;
...
        fscanf(f_lock, "%ld", (long *) &pid);
...
}

This writes an 8-byte long to the location of the 4-byte pid variable,
smashing 4 bytes of the stack.

I doubt this is usefully exploitable (with the stack protector enabled,
it's only overwriting the canary), but it does mean a stale lockfile
causes suck to crash (but helpfully, the lockfile is unlinked before the
crash, so the next run will be successful).

This is obviously an upstream bug, but unfortunately there no longer
appears to be an upstream...